# apps/openapi.yml
openapi: "3.1.0"
info:
  title: Hoomanity API
  version: "1.0.0"
  description: |
    REST API consumed by the Hoomanity frontend.
    All endpoints under `/api/` (except `/api/auth/login`) require a Bearer JWT
    token in the `Authorization` header. A 401 response indicates an expired or
    invalid token; the frontend clears the token and redirects to `/login`.

    Chat replies are asynchronous: `POST /api/chat` returns `202` with an
    `eventId`; the actual reply is delivered via Socket.IO events
    (`chat-result` or `chat-skipped`).
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `POST /api/auth/login`.
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: Admin username.
        password:
          type: string
          description: Admin password.
    LoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: JWT token for authenticated requests.

    # ── Chat ──────────────────────────────────────────────────────────────
...
    MCPOAuthConfig:
      type: object
      required: [redirect_uri]
      properties:
        redirect_uri:
          type: string
        client_id:
          type: string
        client_secret:
          type: string
        scope:
          type: string
        authorization_server_url:
          type: string
    MCPConnectionHosted:
      type: object
      required: [id, type, server_label, server_url]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [hosted]
        server_label:
          type: string
        server_url:
          type: string
        tool_filter:
          type: string
          description: "Comma-separated glob patterns, e.g. *, !send_*. Empty = all."
        headers:
          type: object
          additionalProperties:
            type: string
        oauth:
          $ref: "#/components/schemas/MCPOAuthConfig"
        oauth_has_tokens:
          type: boolean
          description: Read-only; set by server when OAuth tokens exist.
        created_at:
          type: string
          format: date-time
    MCPConnectionStreamableHttp:
      type: object
      required: [id, type, name, url]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [streamable_http]
        name:
          type: string
        url:
          type: string
        tool_filter:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        timeout_seconds:
          type: integer
        cache_tools_list:
          type: boolean
        max_retry_attempts:
          type: integer
        oauth:
          $ref: "#/components/schemas/MCPOAuthConfig"
        oauth_has_tokens:
          type: boolean
        created_at:
          type: string
          format: date-time
    MCPConnectionStdio:
      type: object
      required: [id, type, name, command, args]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [stdio]
        name:
          type: string
        command:
          type: string
        args:
          type: array
          items:
            type: string
        tool_filter:
          type: string
        env:
          type: object
          additionalProperties:
            type: string
        cwd:
          type: string
        created_at:
          type: string
          format: date-time
    MCPConnection:
      oneOf:
        - $ref: "#/components/schemas/MCPConnectionHosted"
        - $ref: "#/components/schemas/MCPConnectionStreamableHttp"
        - $ref: "#/components/schemas/MCPConnectionStdio"
      discriminator:
        propertyName: type
        mapping:
          hosted: "#/components/schemas/MCPConnectionHosted"
          streamable_http: "#/components/schemas/MCPConnectionStreamableHttp"
          stdio: "#/components/schemas/MCPConnectionStdio"
...
    SkillsCommandResponse:
      type: object
      properties:
        output:
          type: string
        error:
          type: string
        code:
          type: integer
          nullable: true

    # ── Audit ─────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      operationId: login
      summary: Authenticate and obtain JWT
      description: |
        Validates username/password against server configuration (argon2 hash)
        and returns a JWT token. No auth header required.

        Returns **501** if web auth is not configured on the server.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Missing username or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: Web auth is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal login failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ╔══════════════════════════════════════════════════════════════════════╗
  # ║  Chat                                                               ║
  # ╚══════════════════════════════════════════════════════════════════════╝
...
  /api/capabilities/mcp/connections:
    get:
      operationId: getMCPConnections
      summary: List all MCP connections
      description: Returns connections with secrets masked by the server.
      tags: [Capabilities]
      responses:
        "200":
          description: MCP connections.
          content:
            application/json:
              schema:
                type: object
                required: [connections]
                properties:
                  connections:
                    type: array
                    items:
                      $ref: "#/components/schemas/MCPConnection"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createMCPConnection
      summary: Add a new MCP connection
      tags: [Capabilities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MCPConnection"
      responses:
        "201":
          description: Connection created (secrets masked in response).
          content:
            application/json:
              schema:
                type: object
                required: [connection]
                properties:
                  connection:
                    $ref: "#/components/schemas/MCPConnection"
        "400":
          description: Invalid connection config.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/capabilities/mcp/connections/{id}:
    patch:
      operationId: updateMCPConnection
      summary: Update an MCP connection
      tags: [Capabilities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MCPConnection"
      responses:
        "200":
          description: Connection updated (secrets masked).
          content:
            application/json:
              schema:
                type: object
                required: [connection]
                properties:
                  connection:
                    $ref: "#/components/schemas/MCPConnection"
        "400":
          description: Invalid patch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Connection not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteMCPConnection
      summary: Delete an MCP connection
      tags: [Capabilities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Connection deleted (no content).
        "404":
          description: Connection not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/capabilities/mcp/oauth/start:
    post:
      operationId: startMCPOAuth
      summary: Start OAuth flow for an MCP connection
      description: |
        Initiates the OAuth authorization flow for a given MCP connection.
        Returns either an `authorizationUrl` to open in the browser, or
        `{ status: "already_authorized" }` if tokens already exist.
      tags: [Capabilities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connectionId]
              properties:
                connectionId:
                  type: string
      responses:
        "200":
          description: OAuth flow initiated or already authorized.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [authorizationUrl]
                    properties:
                      authorizationUrl:
                        type: string
                        format: uri
                  - type: object
                    required: [status]
                    properties:
                      status:
                        type: string
                        enum: [already_authorized]
        "400":
          description: Missing connectionId.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: OAuth start failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ╔══════════════════════════════════════════════════════════════════════╗
  # ║  Skills                                                             ║
  # ╚══════════════════════════════════════════════════════════════════════╝
...
  /api/skills/{id}/content:
    get:
      operationId: getSkillContent
      summary: Get the SKILL.md content for a skill
      tags: [Skills]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skill content.
          content:
            application/json:
              schema:
                type: object
                required: [content]
                properties:
                  content:
                    type: string
        "400":
          description: Missing skill id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Skill not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Read error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/skills/add:
    post:
      operationId: addSkillsPackage
      summary: Install skills from a package
      description: |
        Runs the skills CLI to add a package. Optionally restricts to
        specific skill names within the package.
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [package]
              properties:
                package:
                  type: string
                  description: Package identifier (npm package, git URL, etc.).
                skills:
                  type: array
                  items:
                    type: string
                  description: Optional list of specific skills to install from the package.
      responses:
        "200":
          description: CLI output.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillsCommandResponse"
        "400":
          description: Missing or invalid package.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: CLI execution error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillsCommandResponse"
  /api/skills/remove:
    post:
      operationId: removeSkillsPackage
      summary: Remove installed skills
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skills]
              properties:
                skills:
                  type: array
                  items:
                    type: string
                  description: Skill IDs to remove.
      responses:
        "200":
          description: CLI output.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillsCommandResponse"
        "400":
          description: Missing or invalid skills array.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: CLI execution error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillsCommandResponse"